// src/app/inscriptions/interfaces/pre-student.interface.ts
// Generated by https://quicktype.io

export interface PreStudent {
  id: number;
  name: string;
  dni: string;
  email: string;
  department_id: number;
  address: string;
  phone: string;
  cellphone: string;
  observation: string;
  course_id: number;
  course_turn_id: null;
  signed_up: number;
  registered_by: RegisteredBy;
  registered_at: string;
  enrolled_by: null;
  enrolled_at: null;
  start_date: null;
  created_at: string;
  updated_at: string;
  department: Department;
}

export interface Department {
  id: number;
  name: string;
  created_at: string;
  updated_at: string;
}

export interface RegisteredBy {
  id: number;
  name: string;
  email: string;
  email_verified_at: string;
  created_at: string;
  updated_at: string;
}